refactor: add types, update webSocketMessageHandler.ts
